refactor(acp): extract spawn binding helpers to bindings.ts
